requirements.txt
pygame==2.4.0
numpy==1.24.3
pyyaml==6.0.2
